Move vlms to ilms to make it the name more explicit
